refactor(tui): harden and separate TTY ownership
